+++ CHANGELOG.md
@@ -48,6 +48,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix bug where setting the `<mwc-snackbar>` `labelText` property could throw
   an exception and fail to render
   ([#367](https://github.com/material-components/material-components-web-components/issues/367)).
+- Removed:
+  - mwc-top-app-bar-short has been removed after guidance from the Material Design team.
 
 ## [0.6.0] - 2019-06-05
 - Upgrade lerna to 3.x